Added problem details media type writer

+++ common/src/main/java/dk/sds/nsp/maternity/facade/common/model/ProblemDetails.java
@@ -1,7 +1,7 @@
 package dk.sds.nsp.maternity.facade.common.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -17,7 +17,7 @@ public class ProblemDetails {
     private String title = null;
 
     @JsonProperty("status")
-    private BigDecimal status = null;
+    private Integer status = null;
 
     @JsonProperty("instance")
     private String instance = null;
@@ -67,7 +67,7 @@ private List<ProblemDetailsInvalidparams> invalidParams = null;
         this.title = title;
     }
 
-    public ProblemDetails status(BigDecimal status) {
+    public ProblemDetails status(Integer status) {
         this.status = status;
         return this;
     }
@@ -77,11 +77,11 @@ private List<ProblemDetailsInvalidparams> invalidParams = null;
      *
      * @return status
      **/
-public BigDecimal getStatus() {
+    public Integer getStatus() {
         return status;
     }
 
-    public void setStatus(BigDecimal status) {
+    public void setStatus(Integer status) {
         this.status = status;
     }
